+++ pluslib/Support/Application.php
@@ -29,7 +29,7 @@ class Application extends Container
   /**
    * @var string Pluslib Version
    */
-  const VERSION = '0.0.03';
+  const VERSION = '0.0.04';
 
   function getDefaultBindings()
   {