+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.03...master)
 
+* `Select::first`: apply default value
+
 ## [v0.0.03](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...v0.0.03) - 8/29/2024
 
 * Add `Response` Class