+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Release Notes for 0.x
 
-## [Unreleased](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.03...master)
+## [v0.0.04](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.03...v0.0.0.04)
 
 - `Select::first`: apply default value
 - `Database\DB`: Custom engine (mysql, ...)
 - Implement `ServiceProvider`
 - Use `join_paths` in autoloader
+- Fix phpdoc of `Model::__call`
 
 ## [v0.0.03](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...v0.0.03) - 8/29/2024